Cleanup pathsim split for nosetests
